Slight modification, now compatible with both float and double version of Csound
git-svn-id: http://fs.math.tu-berlin.de:8000/svn/jreality/trunk@4777 6a571d9c-c4e9-67b8-d2a0-cd0d85a3e679

+++ src-audio/de/jreality/audio/csound/CsoundSource.java
@@ -71,11 +71,11 @@ public class CsoundSource extends RingBufferSource {
 			}
 			spout.SetValues(0, bufSize, csnd.GetSpout());
 			for(int j=0; j<ksmps; j++) {
-				double v = 0;
+				float v = 0;
 				for(int k=j; k<bufSize; k+=ksmps) {
 					v += spout.GetValue(k);
 				}
-				ringBuffer.write((float) (v/scale));
+				ringBuffer.write(v/scale);
 			}
 		}
 	}
